LayoutTests/http/tests/misc/resource-timing-resolution-expected.txt

Verifies the minimum resolution is 5 microseconds.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS t0 !== t1 is true
PASS shouldBeNearZeroOrOne < 1e-10 || Math.abs(shouldBeNearZeroOrOne - 1) < 1e-10 is true
PASS successfullyParsed is true

TEST COMPLETE
